cdk-app/requirements.txt
aws-cdk-lib>=2.116.0
constructs>=10.0.0
